Creates reusable functions to generate and display differences
