+++ packages/flow/src/lib/components/RemoteFile.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
 	import { onMount, afterUpdate } from 'svelte';
 	import InkMde from 'ink-mde/svelte';
-	import * as Diff from 'diff';
 	import { marked } from 'marked';
-	import * as Diff2Html from 'diff2html';
 	import 'diff2html/bundles/css/diff2html.min.css';
 	import '../../app.css';
 	import { displayPreview } from '$lib/utils/displayPreview';
+	import { generateDiff } from '$lib/utils/generateDiff';
+	import { displayDiff } from '$lib/utils/displayDiff';
 
 	let apiCall: string = '/api/fetchRemote';
 	let oldText: string = '';
@@ -24,18 +24,6 @@
 		oldText = data;
 		value = data;
 	});
-
-	function displayDiff(): void {
-		isDiffDisplayed = true;
-		const textDiff = Diff.createTwoFilesPatch('file', 'file', oldText, value);
-
-		const diffJson = Diff2Html.parse(textDiff);
-		diffHtml = Diff2Html.html(diffJson, {
-			drawFileList: true,
-			matching: 'lines',
-			outputFormat: 'side-by-side'
-		});
-	}
 
 	afterUpdate(() => {
 		const demo = document.querySelector('#demo');
@@ -79,7 +67,14 @@
 			class="primary-btn"
 			on:click={() => (isPreviewDisplayed = displayPreview(isPreviewDisplayed))}>Preview file</a
 		>
-		<a href="/#diff" class="secondary-btn" on:click={displayDiff}>Display diff</a>
+		<a
+			href="/#diff"
+			class="secondary-btn"
+			on:click={() => {
+				diffHtml = generateDiff(oldText, value);
+				isDiffDisplayed = displayDiff(isDiffDisplayed);
+			}}>Display diff</a
+		>
 	</div>
 	<InkMde
 		bind:value
@@ -111,7 +106,14 @@
 			class="primary-btn"
 			on:click={() => (isPreviewDisplayed = displayPreview(isPreviewDisplayed))}>Preview file</a
 		>
-		<a href="/#diff" class="secondary-btn" on:click={displayDiff}>Display diff</a>
+		<a
+			href="/#diff"
+			class="secondary-btn"
+			on:click={() => {
+				diffHtml = generateDiff(oldText, value);
+				isDiffDisplayed = displayDiff(isDiffDisplayed);
+			}}>Display diff</a
+		>
 	</div>
 
 	{#if isPreviewDisplayed}
