+++ packages/flow/src/lib/components/VSCodeFile.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
 	import { afterUpdate } from 'svelte';
-	import * as Diff from 'diff';
-	import * as Diff2Html from 'diff2html';
 	import InkMde from 'ink-mde/svelte';
 	import TurndownService from 'turndown';
 	import { marked } from 'marked';
 	import 'diff2html/bundles/css/diff2html.min.css';
 	import '../../app.css';
 	import { displayPreview } from '$lib/utils/displayPreview';
+	import { displayDiff } from '$lib/utils/displayDiff';
+	import { generateDiff } from '$lib/utils/generateDiff';
 
 	export let data: { files: { path: string; meta: { title: string } }[] };
 
@@ -21,18 +21,6 @@
 	let isPreviewDisplayed: boolean = false;
 
 	let isDiffDisplayed: boolean = false;
-
-	function displayDiff(): void {
-		isDiffDisplayed = true;
-		const textDiff = Diff.createTwoFilesPatch('file', 'file', oldText, value);
-
-		const diffJson = Diff2Html.parse(textDiff);
-		diffHtml = Diff2Html.html(diffJson, {
-			drawFileList: true,
-			matching: 'lines',
-			outputFormat: 'side-by-side'
-		});
-	}
 
 	async function getFile(fileName: string) {
 		try {
@@ -114,7 +102,14 @@
 				class="primary-btn"
 				on:click={() => (isPreviewDisplayed = displayPreview(isPreviewDisplayed))}>Preview file</a
 			>
-			<a href="/#diff" class="secondary-btn" on:click={displayDiff}>Display diff</a>
+			<a
+				href="/#diff"
+				class="secondary-btn"
+				on:click={() => {
+					diffHtml = generateDiff(oldText, value);
+					isDiffDisplayed = displayDiff(isDiffDisplayed);
+				}}>Display diff</a
+			>
 		</div>
 		<InkMde
 			bind:value
@@ -147,7 +142,14 @@
 			class="primary-btn"
 			on:click={() => (isPreviewDisplayed = displayPreview(isPreviewDisplayed))}>Preview file</a
 		>
-		<a href="/#diff" class="secondary-btn" on:click={displayDiff}>Display diff</a>
+		<a
+			href="/#diff"
+			class="secondary-btn"
+			on:click={() => {
+				diffHtml = generateDiff(oldText, value);
+				isDiffDisplayed = displayDiff(isDiffDisplayed);
+			}}>Display diff</a
+		>
 	</div>
 
 	{#if isPreviewDisplayed}
