+++ packages/flow/src/lib/utils/displayDiff.ts
@@ -1,0 +1,5 @@
+export function displayDiff(isDiffDisplayed: boolean): boolean {
+	console.log('Diff is displayed');
+	isDiffDisplayed = true;
+	return isDiffDisplayed;
+}
